+++ src/plugins/data/server/search/opensearch_search/decide_client.ts
@@ -9,14 +9,13 @@ import { IOpenSearchSearchRequest } from '..';
 export const decideClient = async (
   context: RequestHandlerContext,
   request: IOpenSearchSearchRequest,
-  withDataSourceEnabled: boolean = false,
   withLongNumeralsSupport: boolean = false
 ): Promise<OpenSearchClient> => {
   const defaultOpenSearchClient = withLongNumeralsSupport
     ? context.core.opensearch.client.asCurrentUserWithLongNumeralsSupport
     : context.core.opensearch.client.asCurrentUser;
 
-  return withDataSourceEnabled && request.dataSourceId && context.dataSource
+  return request.dataSourceId && context.dataSource
     ? await context.dataSource.opensearch.getClient(request.dataSourceId)
     : defaultOpenSearchClient;
 };