+++ src/plugins/data/server/search/opensearch_search/opensearch_search_strategy.ts
@@ -81,12 +81,7 @@ export const opensearchSearchStrategyProvider = (
           throw new Error(`Data source id is required when no openseach hosts config provided`);
         }
 
-        const client = await decideClient(
-          context,
-          request,
-          dataSource?.dataSourceEnabled(),
-          withLongNumeralsSupport
-        );
+        const client = await decideClient(context, request, withLongNumeralsSupport);
         const promise = shimAbortSignal(client.search(params), options?.abortSignal);
 
         const { body: rawResponse } = (await promise) as ApiResponse<SearchResponse<any>>;